Add support for @event, @listens, @fires.
@event: Rendered at the top of each documentation page as "Events" above
"Methods".

@listens: Rendered at the top of each documentation page as "Event
Listeners" above "Methods" if the @listens tag was declared at the class
level. If @listens is declared at a method level an "Event Listeners"
section is rendered under the "return" documentation section for that
method.

@fires: Rendered at the top of each documentation page as "Events
Triggered" above "Methods" if the @fires tag was declared at the class
level. If @fires is declared at a method level an "Events Triggered"
section is rendered under the "return" documentation section for that
method.